+++ src/MEDYAN.cpp
@@ -84,7 +84,44 @@ int main(int argc, char **argv) {
     cout << "         of Active Networks, Third Generation.         " << endl;
     cout << "         PAPOIAN LAB 2015, ALL RIGHTS RESERVED         " << endl;
     cout << "*******************************************************" << endl;
-    
+    cout << "MEDYAN version:                      v4.2.1(unreleased)"<<endl;
+
+    cout << "Memory model:                        "<< static_cast<unsigned>(8 * sizeof
+    (void*))<<" bit"<<endl;
+
+    cout << "Coordinate/Force precision:          ";
+    #if FLOAT_PRECISION
+    cout << "single" << endl;
+    #else
+    cout << "double" << endl;
+    #endif
+
+    cout << "Pair-wise list algorithm:            ";
+    #if defined NLORIGINAL
+    cout << "original (legacy)"<<endl;
+    #elif defined NLSTENCILLIST
+    cout << "stencil list based"<<endl;
+    #elif defined HYBRID_NLSTENCILLIST
+    cout << "hybrid**"<<endl;
+    cout << "**single neighbor list for compatible distance ranges corresponding to the "
+            "same filament type pairs"<<endl;
+    #elif defined SIMDBINDINGSEARCH
+    cout << "SIMD**"<<endl;
+    cout << "SIMD instruction set:                ";
+    #if defined __AVX512F__
+    cout<<"AVX512 (MEDYAN will use AVX2 instructions"<<endl;
+    #elif defined __AVX2__
+    cout<<"AVX2"<<endl;
+    #elif defined __AVX__
+    cout<<"AVX"<<endl;
+    #else
+    cout<<"none"<<endl;
+    #endif
+    cout<<"**SIMD accelerated hybrid search - single neighbor list for compatible distance"
+          " ranges corresponding to the same filament type pairs"<<endl;
+    #endif
+    cout << "*******************************************************" << endl;
+
     cout.precision(8);
 
     auto cmdRes = medyanInitFromCommandLine(argc, argv);
@@ -118,4 +155,3 @@ int main(int argc, char **argv) {
 
     return returnCode;
 }
-