+++ src/Mechanics/Minimizer/CGPolakRibiereMethod.cpp
@@ -90,6 +90,12 @@ MinimizationResult PolakRibiere::minimize(ForceFieldManager &FFM, floatingpoint 
 #ifdef ALLSYNC
     cudaDeviceSynchronize();
 #endif
+    #ifdef OPTIMOUT
+    cout<<"Energy before minimization"<<endl;
+    FFM.computeEnergy(Bead::getDbData().coords.data(), true);
+    CUDAcommon::tmin.computeenerycallszero++;
+    cout<<endl;
+    #endif
     //@@@{ STEP 2: COMPUTE FORCES
     tbegin = chrono::high_resolution_clock::now();
     FFM.computeForces(Bead::getDbData().coords.data(),
@@ -133,12 +139,11 @@ MinimizationResult PolakRibiere::minimize(ForceFieldManager &FFM, floatingpoint 
     Bead::getDbData().forcesAux = Bead::getDbData().forces;
     Bead::getDbData().forcesAuxP = Bead::getDbData().forces;
     auto maxForce = maxF();
-
-    result.energiesBefore = FFM.computeEnergyHRMD(Bead::getDbData().coords.data());
-
+    cout<<"maxForce "<<maxForce<<endl;
     tend = chrono::high_resolution_clock::now();
     chrono::duration<floatingpoint> elapsed_copy(tend - tbegin);
     CUDAcommon::tmin.copyforces += elapsed_copy.count();
+    result.energiesBefore = FFM.computeEnergyHRMD(Bead::getDbData().coords.data());
     //@@@}
 #endif
     //M as the first letter in variables signifies that it is used by minimizer
@@ -968,9 +973,7 @@ MinimizationResult PolakRibiere::minimize(ForceFieldManager &FFM, floatingpoint 
                 << numIter << " steps." << endl;
         cout << "Maximum force in system = " << maxF() << endl;
         cout << "System energy..." << endl;
-        #ifdef ADDITIONALINFO
         FFM.computeEnergy(Bead::getDbData().coords.data(), true);
-        #endif
     }
 
     else if (numIter >= N) {