+++ src/TESTS/Mechanics/ForceField/Branching/TestBranchingDihedralQuadratic.cpp
@@ -55,6 +55,14 @@ TEST_CASE("Force field: Branching Dihedral Quadratic", "[ForceField]") {
     pos.push_back(0.4);
     kdih.push_back(1.0);
 
+    // special case dihedral angle
+    coords.push_back(V3 { 0.0, -1.0, 0.0 });
+    coords.push_back(V3 { 0.0, 1.0, 0.0 });
+    coords.push_back(V3 { 1.0, 0.0, 0.0 });
+    coords.push_back(V3 { 1.5, -1.0, 0.1 });
+    pos.push_back(1.0);
+    kdih.push_back(1.0);
+
     const auto nint = kdih.size();
     vector< unsigned > beadSet(coords.size());
     std::iota(beadSet.begin(), beadSet.end(), 0u);