+++ src/TESTS/Mechanics/ForceField/Branching/TestBranchingPositionCosine.cpp
@@ -4,12 +4,12 @@
 
 #include "catch2/catch.hpp"
 
-#include "Mechanics/ForceField/Branching/BranchingDihedralCosine.h"
+#include "Mechanics/ForceField/Branching/BranchingPositionCosine.h"
 #include "TESTS/Mechanics/ForceField/TestFFCommon.hpp"
 
 using namespace test_ff_common;
 
-TEST_CASE("Force field: Branching Dihedral Cosine", "[ForceField]") {
+TEST_CASE("Force field: Branching Position Cosine", "[ForceField]") {
     // The test case checks whether energy and force are consistent
 
     using namespace std;
@@ -19,64 +19,59 @@ TEST_CASE("Force field: Branching Dihedral Cosine", "[ForceField]") {
 
     // Prepare data
     //---------------------------------
-    vector< floatingpoint > kdih;
+    vector< floatingpoint > kpos;
     vector< floatingpoint > pos;
     VA3 coords;
 
-    // small dihedral angle
+    //
     coords.push_back(V3 { 0.0, -1.0, 0.0 });
     coords.push_back(V3 { 0.0, 1.0, 0.0 });
     coords.push_back(V3 { 1.0, 0.0, 0.0 });
-    coords.push_back(V3 { 1.5, 1.0, 0.1 });
     pos.push_back(0.6);
-    kdih.push_back(1.0);
+    kpos.push_back(1.0);
 
     // 90 deg dihedral angle (<90)
     coords.push_back(V3 { 0.0, -1.0, 0.0 });
     coords.push_back(V3 { 0.0, 1.0, 0.0 });
     coords.push_back(V3 { 1.0, 0.0, 0.0 });
-    coords.push_back(V3 { 1.5, 0.1, 1.0 });
     pos.push_back(0.5);
-    kdih.push_back(1.0);
+    kpos.push_back(1.0);
 
     // 90 deg dihedral angle (>90)
     coords.push_back(V3 { 0.0, -1.0, 0.0 });
     coords.push_back(V3 { 0.0, 1.0, 0.0 });
     coords.push_back(V3 { 1.0, 0.0, 0.0 });
-    coords.push_back(V3 { 1.5, -0.1, 1.0 });
     pos.push_back(0.5);
-    kdih.push_back(1.0);
+    kpos.push_back(1.0);
 
     // large dihedral angle
     coords.push_back(V3 { 0.0, -1.0, 0.0 });
     coords.push_back(V3 { 0.0, 1.0, 0.0 });
     coords.push_back(V3 { 1.0, 0.0, 0.0 });
-    coords.push_back(V3 { 1.5, -1.0, 0.1 });
     pos.push_back(0.4);
-    kdih.push_back(1.0);
+    kpos.push_back(1.0);
 
     // special case dihedral angle
     coords.push_back(V3 { 0.0, -1.0, 0.0 });
     coords.push_back(V3 { 0.0, 1.0, 0.0 });
     coords.push_back(V3 { 1.0, 0.0, 0.0 });
-    coords.push_back(V3 { 1.5, -1.0, 0.1 });
     pos.push_back(1.0);
-    kdih.push_back(1.0);
+    kpos.push_back(1.0);
 
-    const auto nint = kdih.size();
+    const auto nint = kpos.size();
     vector< unsigned > beadSet(coords.size());
     std::iota(beadSet.begin(), beadSet.end(), 0u);
 
     // Prepare functions
     //---------------------------------
     const auto calcEnergy = [&](const VA3& c) {
-        return BranchingDihedralCosine {}.energy(
-            c.data(), nint, beadSet.data(), kdih.data(), pos.data()
+        return BranchingPositionCosine {}.energy(
+            c.data(), beadSet.data(), kpos.data(), pos.data()
         );
     };
     const auto calcForce = [&](const VA3& c, VA3& f) {
-        BranchingDihedralCosine {}.forces(
-            c.data(), f.data(), nint, beadSet.data(), kdih.data(), pos.data(), nullptr
+        BranchingPositionCosine {}.forces(
+            c.data(), f.data(), beadSet.data(), kpos.data(), pos.data(), nullptr
         );
     };